+++ chipset/da145xx/btstack_chipset_da145xx.h
@@ -36,14 +36,12 @@
  */
 
 /*
- *  btstack_chipset_da14581.c
- *
  *  Adapter to use da14581-based chipsets with BTstack
  *  
  */
 
-#ifndef BTSTACK_CHIPSET_DA14581_H
-#define BTSTACK_CHIPSET_DA14581_H
+#ifndef BTSTACK_CHIPSET_DA145XX_H
+#define BTSTACK_CHIPSET_DA145XX_H
 
 #if defined __cplusplus
 extern "C" {
@@ -55,22 +53,22 @@ extern "C" {
 /**
  * @brief get chipset instance
  */
-const btstack_chipset_t * btstack_chipset_da14581_instance(void);
+const btstack_chipset_t * btstack_chipset_da145xx_instance(void);
 
 /**
  * @brief Download firmware via btstack_uart_t implementation
  * @param uart_driver -- already initialized
  * @param done callback. 0 = Success
  */
-void btstack_chipset_da14581_download_firmware_with_uart(const btstack_uart_t * uart_driver, const uint8_t * fw, uint16_t fw_size, void (*done)(int result));
+void btstack_chipset_da145xx_download_firmware_with_uart(const btstack_uart_t * uart_driver, const uint8_t * fw, uint16_t fw_size, void (*done)(int result));
 
 /**
  * @brief Download firmware via btstack_uart_block_t implementation
  * @param uart_driver -- already initialized
  * @param done callback. 0 = Success
- * @deprecated please use btstack_chipset_da14581_download_firmware_with_uart instead
+ * @deprecated please use btstack_chipset_da145xx_download_firmware_with_uart instead
  */
-void btstack_chipset_da14581_download_firmware(const btstack_uart_block_t * uart_driver, const uint8_t * fw, uint16_t fw_size, void (*done)(int result));
+void btstack_chipset_da145xx_download_firmware(const btstack_uart_block_t * uart_driver, const uint8_t * fw, uint16_t fw_size, void (*done)(int result));
 
 #if defined __cplusplus
 }